Refactor package names to be called 'configuration' instead of 'descriptor'.
