+++ src/main/java/com/synopsys/integration/alert/database/api/configuration/ConfigContextEnum.java
@@ -1,0 +1,6 @@
+package com.synopsys.integration.alert.database.api.configuration;
+
+public enum ConfigContextEnum {
+    GLOBAL,
+    DISTRIBUTION
+}
