--- src/main/java/com/synopsys/integration/alert/database/api/descriptor/ConfigContextEnum.java
@@ -1,6 +1,0 @@
-package com.synopsys.integration.alert.database.api.descriptor;
-
-public enum ConfigContextEnum {
-    GLOBAL,
-    DISTRIBUTION
-}
