+++ src/main/java/com/synopsys/integration/alert/database/repository/configuration/DescriptorFieldRepository.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.alert.database.repository.descriptor;
+package com.synopsys.integration.alert.database.repository.configuration;
 
 import java.util.List;
