+++ src/test/java/com/synopsys/integration/alert/repository/configuration/DescriptorConfigRepositoryTestIT.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.alert.repository.descriptor;
+package com.synopsys.integration.alert.repository.configuration;
 
 import static org.junit.Assert.assertEquals;
 
@@ -12,9 +12,9 @@ import com.synopsys.integration.alert.AlertIntegrationTest;
 import com.synopsys.integration.alert.database.entity.descriptor.ConfigContextEntity;
 import com.synopsys.integration.alert.database.entity.descriptor.DescriptorConfigEntity;
 import com.synopsys.integration.alert.database.entity.descriptor.RegisteredDescriptorEntity;
-import com.synopsys.integration.alert.database.repository.descriptor.ConfigContextRepository;
-import com.synopsys.integration.alert.database.repository.descriptor.DescriptorConfigRepository;
-import com.synopsys.integration.alert.database.repository.descriptor.RegisteredDescriptorRepository;
+import com.synopsys.integration.alert.database.repository.configuration.ConfigContextRepository;
+import com.synopsys.integration.alert.database.repository.configuration.DescriptorConfigRepository;
+import com.synopsys.integration.alert.database.repository.configuration.RegisteredDescriptorRepository;
 
 public class DescriptorConfigRepositoryTestIT extends AlertIntegrationTest {
     public static final String DESCRIPTOR_NAME = "Test Descriptor";
